Initial setup of EKS, VPC, and Role Groupings
This commit still request a change to the generated modules/eks/local.tf  root volume type from gp3->gp2.

A further commit will look to address this. I assume at the time of committing this is due to me using a later version of module that is setting gp3 as default. 

What is a gp3? Still learning...